Made it possible to "register" computers as "energy computers" and for those computers to send data to the database.
Made endpoints to the API that handles POST and GET requests to "api/energyComputer".

This makes it now possible to register a CC computer by sending a json string that include the computers ID, the max energy, in RF, that the connected energy storage device can hold and the computers label, can be an empty string.

It also makes it possible to get computers by sending a GET request with optional parameters "numComputers" and "computerID".

Updated sendEnergyData function to include "computerID" as a parameter.

Added the response "Online" when sending a GET request to the "api/" endpoint.

Changed checkError() to handleError() in order to make it the same across files.